ADD: ItadItemInfo into a pseudo component
